requirements.txt
python-dotenv
discord.py
grequests
SQLAlchemy

# For postgresql
psycopg2
# For sqlite
aiosqlite